refactor(transaction service): set right ubication of components
