+++ transaction-service/src/domain/transaction.service.ts
@@ -1,10 +1,10 @@
 import { Inject, Injectable, Logger } from '@nestjs/common';
 import { ClientKafka } from '@nestjs/microservices';
 import { InjectRepository } from '@nestjs/typeorm';
-import { CreateTransactionDto } from 'src/domain/create-transaction.dto';
-import { UpdateTransactionDto } from 'src/domain/update-transaction.dto';
 import { Repository } from 'typeorm';
-import { Transaction } from '../domain/transaction.entity';
+import { Transaction } from './transaction.entity';
+import { CreateTransactionDto } from './dto/create-transaction.dto';
+import { UpdateTransactionDto } from './dto/update-transaction.dto';
 
 @Injectable()
 export class TransactionService {
